Move View\Error to View\Exception.
