+++ src/Error/ExceptionRenderer.php
@@ -25,7 +25,7 @@ use Cake\Network\Request;
 use Cake\Network\Response;
 use Cake\Routing\Router;
 use Cake\Utility\Inflector;
-use Cake\View\Error\MissingViewException;
+use Cake\View\Exception\MissingViewException;
 use Exception;
 
 /**
