+++ src/View/Exception/MissingCellException.php
@@ -4,27 +4,24 @@
  * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  *
  * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
  * Redistributions of files must retain the above copyright notice.
  *
  * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- * @link          http://book.cakephp.org/2.0/en/development/testing.html
+ * @link          http://cakephp.org CakePHP(tm) Project
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\View\Error;
+namespace Cake\View\Exception;
 
 use Cake\Core\Exception\Exception;
 
 /**
- * Used when an element file cannot be found.
- */
-class MissingElementException extends Exception {
-
-/**
- * Message template
+ * Used when a cell class file cannot be found.
  *
- * @var string
  */
-	protected $_messageTemplate = 'Element file "%s" is missing.';
+class MissingCellException extends Exception {
+
+	protected $_messageTemplate = 'Cell class "%s" is missing.';
 
 }
