+++ src/View/Exception/MissingViewException.php
@@ -4,24 +4,23 @@
  * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  *
  * Licensed under The MIT License
- * For full copyright and license information, please see the LICENSE.txt
  * Redistributions of files must retain the above copyright notice.
  *
  * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- * @link          http://cakephp.org CakePHP(tm) Project
+ * @link          http://book.cakephp.org/2.0/en/development/testing.html
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\View\Error;
+namespace Cake\View\Exception;
 
 use Cake\Core\Exception\Exception;
 
 /**
- * Used when a cell class file cannot be found.
+ * Used when a view file cannot be found.
  *
  */
-class MissingCellException extends Exception {
+class MissingViewException extends Exception {
 
-	protected $_messageTemplate = 'Cell class "%s" is missing.';
+	protected $_messageTemplate = 'View file "%s" is missing.';
 
 }
