+++ tests/TestCase/Controller/PagesControllerTest.php
@@ -21,7 +21,7 @@ use Cake\Core\Configure;
 use Cake\Network\Request;
 use Cake\Network\Response;
 use Cake\TestSuite\TestCase;
-use Cake\View\Error\MissingViewException;
+use Cake\View\Exception\MissingViewException;
 use TestApp\Controller\PagesController;
 
 /**
@@ -60,7 +60,7 @@ class PagesControllerTest extends TestCase {
 /**
  * Test that missing view in debug mode renders missing_view error page
  *
- * @expectedException \Cake\View\Error\MissingViewException
+ * @expectedException \Cake\View\Exception\MissingViewException
  * @expectedExceptionCode 500
  * @return void
  */
