+++ tests/TestCase/Error/ExceptionRendererTest.php
@@ -37,9 +37,9 @@ use Cake\Network\Request;
 use Cake\ORM\Error\MissingBehaviorException;
 use Cake\Routing\Router;
 use Cake\TestSuite\TestCase;
-use Cake\View\Error\MissingHelperException;
-use Cake\View\Error\MissingLayoutException;
-use Cake\View\Error\MissingViewException;
+use Cake\View\Exception\MissingHelperException;
+use Cake\View\Exception\MissingLayoutException;
+use Cake\View\Exception\MissingViewException;
 
 /**
  * BlueberryComponent class
